move all tests outside firebase to one folder
