+++ raheem/src/components/Testing/map.spec.js
@@ -1,10 +1,15 @@
 import React from 'react';
 import { render, fireEvent } from '@testing-library/react';
+import ReactMapGL, { GeolocateControl } from 'react-map-gl';
 
 import Map from '../Dashboard/Map';
 
+// const TOKEN = process.env.REACT_APP_MAPBOX_ACCESSTOKEN;
+
+
 describe('<Map />', () => {
-    it('renders successfully', () => {
+    it('renders without crashing', () => {
         render(<Map />);
     });
 });
+
